Rename components directories and make them lowercase
