+++ src/App.tsx
@@ -2,11 +2,11 @@ import React from "react";
 import "./variables.css";
 import "./colorscheme.css";
 import AlertView from "./views/alert/AlertView";
-import LoginView from "./views/loginView/LoginView";
+import LoginView from "./views/login/LoginView";
 import { Route, Switch, useHistory } from "react-router-dom";
 import { ProtectedRoute } from "./protected.route";
-import NotificationProfileView from "./views/notificationProfileView/NotificationProfileView";
-import FilterBuildingView from "./views/filterBuildingView/filterBuildingView";
+import NotificationProfileView from "./views/notificationprofile/NotificationProfileView";
+import FilterBuildingView from "./views/filterbuilding/filterBuildingView";
 import TimeslotView from "./views/timeslot/TimeslotView";
 
 import api from "./api";
