+++ src/views/filterbuilding/filterBuildingView.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Header from "../../components/header/Header";
-import "../../components/react-table/table.css";
-import FilterBuilder from "../../components/filterBuilder/filterBuilder";
+import "../../components/alerttable/alerttable.css";
+import FilterBuilder from "../../components/filterbuilder/filterBuilder";
 import { withRouter } from "react-router-dom";
 type PropType = {
   history: any;
